+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - Added `display_name` property to `Task` (e.g. for plotting).
+- Analysis: `task_info()` operation for data frame preparation.
 
 ## 0.3.115 (26 July 2025)